Fix WorkStealingDeque (#31)
`WorkStealingDeque{Any}` sometimes threw `UndefRefError: access to
undefined reference` before this patch.  This was because Julia tries
to "verify" `buffer[top]` at the time it is invoked. However, we can
only do so after ensuring that `buffer[top]` is a valid element upon
the CAS.  This patch fixes the issue by using
`unsafe_pointer_to_objref` to delay materializing the Julia object
until the CAS.