Redesign dictionary: use Vector{DictSlot{K,V}} (#39)
Since Julia does not have atomics for arrays, supporting boxed Julia
objects in a hash table requires indirection (i.e., a `mutable
struct`) *somewhere* at the moment because no atomic operation can be
defined on `Vector{Any}`. Previously, `LinearProbingDict` was using a
`mutable struct` only for "storage" and the actual linearization was
done on a dense array using 128 bit CAS (based on Maier et
al. (2019)). However, it required complex key/value storage management
to actually store objects in the hash table. It turned out that just
using `Vector{DictSlot{K,V}}` (where `DictSlot{K,V}` is a `mutable
struct`) is more straightforward and beneficial in terms of
performance. The important optimization to make it practical is to
avoid copying `DictSlot{K,V}` during migration.

A nice side-effect of this approach is that it gets rid of the "torn"
64-bit load which is not likely to be supported in Julia. It is now
possible to separate key and value to distinct `Vector`s if atomics on
`Vector{Any}` is supported at some point. This is more likely to be
supported than updating key/value location directly in
`Vector{Pair{Any,Any}}`.